add age calculation and daily calorie management features

+++ app/src/main/java/com/duzceders/aicaltracker/product/models/User.java
@@ -19,15 +19,15 @@ public class User {
     private int height_cm;
     private int weight_kg;
     private int target_weight_kg;
-    private double body_fat_percent;
+    private int body_fat_percent;
 
     private ActivityLevel activity_level;
 
-    private double bmr;
-    private double daily_calorie_needs;
-    private double daily_calorie_needs_left;
-    private double daily_water_needs_liters;
-    private double daily_water_needs_left_liters;
+    private int bmr;
+    private int daily_calorie_needs;
+    private int daily_calorie_needs_left;
+    private int daily_water_needs_liters;
+    private int daily_water_needs_left_liters;
 
     private CalorieGoals calorie_goals;
     private DailyMacros daily_macros;
@@ -40,9 +40,9 @@ public class User {
     @Setter
     public static class CalorieGoals {
 
-        private double maintain;
-        private double weight_gain;
-        private double weight_loss;
+        private int maintain;
+        private int weight_gain;
+        private int weight_loss;
 
         public CalorieGoals() {
         }
@@ -52,13 +52,13 @@ public class User {
     @Getter
     @Setter
     public static class DailyMacros {
-        private double daily_carbs_need_g;
-        private double daily_fats_need_g;
-        private double daily_proteins_need_g;
+        private int daily_carbs_need_g;
+        private int daily_fats_need_g;
+        private int daily_proteins_need_g;
 
-        private double daily_carbs_left_g;
-        private double daily_fats_left_g;
-        private double daily_proteins_left_g;
+        private int daily_carbs_left_g;
+        private int daily_fats_left_g;
+        private int daily_proteins_left_g;
 
         public DailyMacros() {
         }
